+++ src/features/payments/components/PayPrivatelyButton.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from '@/components/ui/Button';
+import Button from '@/ui/primitives/Button';
 import { chainAdapter } from '@/services/chain';
 
 interface PayPrivatelyButtonProps {
